linux-imx: Fix inline error [YOCIMX-3640]
Meet kodi compiling error based on thud 4.19.35-1.0.0:
/opt/bamboo/build/kodi/bld-kodi/tmp/work/imx6qpsabresd-poky-linux-gnueabi/kodi-imx/17-r1.0/recipe-sysroot/usr/include/c++/8.2.0/arm-poky-linux-gnueabi/bits/c++config.h:260:10: error: expected unqualified-id before 'namespace'
   inline namespace __cxx11 __attribute__((__abi_tag__ ("cxx11"))) { }
          ^~~~~~~~~
.....

Macros 'inline' and '__gnu_inline' used to be defined in compiler-gcc.h,
which was (and is) included entirely in (__KERNEL__ && !__ASSEMBLY__).
Commit 815f0dd ("include/linux/compiler*.h: make compiler-*.h mutually
exclusive") had those macros exposed to userspace, unintentionally.

Other macros like __must_check, notrace, etc. are in a similar situation.
So just move all these macros back into (__KERNEL__ && !__ASSEMBLY__).

This fix is similar to an upstream fix but adapted due to out-of-sync codebase.
[adapted from upstream: https://github.com/torvalds/linux/commit/71391bdd2e9aab188f86bf1ecd9b232531ec7eea#diff-43ac6627cd65b95a3f7d79ea81e5ff17]

Signed-off-by: Yuqing Zhu <carol.zhu@nxp.com>